Refactor: - routes - rewrite some logic
